Add identifier sets
This patch adds support for sets of the numeric identifiers that we'll
use to identify events and processes in a CSP script.  There are two
separate types: a write-only "set builder", and a read-only "set".
We're currently using Judy arrays in the writeable builder, letting it
take care of sorting and deduplication.  When you build into the
readable set type, we persist that as a sorted linear array.