Keep track of event names
CSP events will usually be represented by a number, but also have unique
names.  These new functions let you add events to a CSP environment and
get their names.

We're going to use hashes for event and process IDs, in the hope that
being reproducible will make it easier to eventually scale up to
multicore refinement checks.  That means that the "create an event"
function is actually "get the ID for an event", since you can call it
multiple times with the same event name and you'll get the same ID.