Add helper functions for constructing process IDs
We want to ensure that process IDs are reproducible, and only depend on
the definition of the process.  (If you try to define two processes with
exactly the same definition, you should get the same ID for both,
without having to coordinate with anyone.)

To that end, this patch adds some helper functions for constructing IDs.
All IDs start from a "scope", which is just a distinct nonce that each
kind of process operator can use to start generating IDs from a distinct
basis.  There are then additional helper functions for adding individual
IDs, sets of IDs, and string names into an existing ID.  Together, this
should give us all of the pieces needed for all of the CSP operators.