Don't rely on constructor order in test cases
Before we assumed that constructor functions, which we use to register
each test case in a test program, would be called in the order that
they're defined in the source file.  That's not guaranteed, however, and
GCC 4.8 changed the ordering! [1]  So now we have to explicitly sort the
list of test cases (using the line number where they're defined as the
sorting key) before calling them.

Note that this is just a cosmetic change, in that it doesn't change
whether any particular test case succeeds or fails; it just changes
which order they're called in, so that the test program output lines up
with what's defined in the source file.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=46770