Create new references to after processes
When getting the `afters` for a particular process and event, we now
bump the reference count of the results.  We need this because we'll
have some operators that create the after processes lazily, and we need
to make sure that the newly created processes aren't immediately freed
and aren't leaked.