Fix memory leak in choice operators
We translate A □ B into □ {A,B}.  This usually isn't a problem, but if A
== B, this actually turns into □ {A}.  We're given two references to
this process (one via A and one via B), but by collapsing this into a
singleton set, we lose one of those references and never free it in our
destructor.  To get around this we now detect the identical processes in
the constructor and preemptively free one of the refernces.

(All of this applies to ⊓, too.)