Add nicer macros for testing initials and afters
This relies on some minor preprocessor magic, but makes it MUCH nicer to
test the initials and afters sets of CSP₀ processes.  Both macros work
when the expected result is empty or non-empty.  For afters sets, you
provide the expected values as CSP₀ process descriptions, and we
automatically parse that for you.