Clean up CSP₀ test macros
The helper macros that verify the "initials" and "afters" sets of a CSP₀
process now require you to use an extra set of parentheses around the
expected contents of the set.  This is most useful in the "afters"
macro, since it makes it clearer which parameter is the `initial` event
and which parameters are the expected `afters` processes.  It also makes
it clearer when we expect the set to be empty, since you see a nice
glaring `()` in the test case.

As part of this, I've also moved all of the test macros that were only
used in one test case to live inside of that test case, to reduce the
clutter in test-cases.h.