Rework process initialization
The CSP environment is now responsible for allocating and freeing the
userdata of each process.  The process interface has two new callbacks:
`get_id` (which returns the ID of a process given its definition), and
`init_ud` (which fills in the "permanent" userdata given a "temporary"
userdata).  This makes it a little bit simpler to create processes, and
reduces the memory churn when defining a process multiple times.