Add event closure function
This new function calculates the closure of a set of CSP processes for a
particular event.  The closure is the set of processes that can be
reached from any of the initial processes by following any sequence of
that event.  (That includes the empty sequence, which means that all of
the initial processes are in the closure.) When we start implementation
CSP normalization, we'll be looking specifically for τ closures.

The pseudo-code algorithm is:

  CLOSURE(initial_nodes, event): closure_nodes
    seen = ∅
    closure_nodes = ∅
    queue = initial_nodes
    for x ∈ queue:
      seen ∪= {x}
      closure_nodes ∪= {x}
      queue ∖= {x}
      for y ∈ afters(x, event):
        if y ∉ seen:
          queue ∪= {y}