Add behavior type
The new csp_behavior type is used to describe the behavior of a process
in a particular semantic model.  Right now we only support one model —
traces — but this type should make it easier to add support for the
other models without having to make deep changes to the refinement code.
All of the bits that are particular to each semantic model will be
hidden away in this new type and its related functions.