Add recursive processes
You can now define mutually recursive processes.  In CSP₀, this is done
via the `let/within` statement.  (This has the same form as in CSPM, but
without allowing you to parameterize the process definitions.)  Names
are only in scope within the specific `let` statement where they're
defined; we don't (currently) allow you to refer to processes in nested
`let` statements.