Add debug recursion identifiers
You can now use the `X@0` syntax to refer to a recursive process from
outside of its `let` definition.  This should only be used during test
cases; it allows us to "crack open" the let statement so that we can
identify and verify the processes defined inside of it.  The number
after the `@` is the "scope identifier" of the recursion; right now we
rely on the fact that each CSP environment automatically assigns
sequential IDs to each `let` statement.