Bisimulate normalized LTSes
We can now bisimulate a normalized LTS, which is part of the larger
refinement algorithm described in Roscoe's book.  Bisimulation is
similar to our existing csp_behavior type, in that it tries to determine
if two processes have equivalent behavior.  The difference is that
csp_behavior only models a process's *immediate* behavior (for instance,
in the traces model, its set of initial events).  The bisimulation
relation takes into account everything "downstream" of the process, as
well.  Two nodes that are equivalent under bisimulation should be
completely indistinguishable from each other under the semantic model in
question.