Fix flaky bisimulation behavior
Before we updated a single set of equivalence classes in place during
each iteration of the bisimulation loop.  That's incorrect, because we
need to check whether each pair was equivalent _as of when the previous
iteration finished_.  If we move a node into a new equivalence class, we
should only pick that up in our comparisons in the _next_ iteration.