Don't install libhst
Let's just focus on the command-line hst tool (which doesn't have
anything inside of it yet...)  We still need to define it as a
"convenience library", so that we can link our test cases with it too;
we just don't want to install it.  This has the side effect of compiling
and linking it statically.