Simplify process interface
Jeez that was way too complex.  Each operator has its own type (as
before), but now with a copy of `struct csp_process` embedded inside of
them.  I've also thrown away the whole "temporary userdata" nonsense.
Each operator constructor now explicitly checks to see whether there's
already a process with the same ID, returning it directly if so.  If the
process is new, we allocate it using a perfectly standard malloc(3)
call.