Use visitor to handle process's afters
Just like we just did for the `initials` method, you now pass in a
visitor to a process's `afters` method instead of a set.  And just like
with the `initials`, the legacy function for grabbing the afters via a
process ID still renders everything into a set, allowing us to migrate
things one by one.