Find all reachable subprocesses
This patch adds a helper function that visits all of the subprocesses
reachable from a given root process.  We ensure that your visitor is
called exactly once for each subprocess.

I've also added a new assertion to each of the CSP operator test cases,
which validate the reachable subprocesses of each test process.