Use same process interface for normalization
Before we the result of normalizing a process as an LTS, using its own
type.  I realized later than I should that a normalized process isn't
different than any other, except for what kinds of edges it's supposed
to have.  So now we use the same `csp_process` "class" to represent
prenormalized and normalized processes.