Keep track of process names
The process interface now includes a method for rendering the name of a
process.  Or maybe their definitions?  We compute the names on the fly
instead of storing them, which should help with memory consumption on
especially large processes.  The process interface also includes the
precedence level of the operator in question, which lets us decide
whether we need to include parentheses when rendering the name of a
bunch of nested subprocesses.

We don't have a good implementation of this new method for recursive
process definitions yet.  I have a thought on how to do it, but it will
be more complex than the implementations for the other operators.