Calculate name of recursive processes
And this patch adds a proper `name` implementation for recursive
processes.  When we encounter a recursive definition, we do a quick BFS
to find all of the other definitions that are mutually recursive with
it.  We then output a `let` statement containing all of those processes.
(This means that we don't necessarily include all of the processes that
were in the original `let` statement, if any of them weren't actually
used.)