Fix interleave's handling of duplicate processes
P ⫴ P is not the same thing as ⫴{P} or P!  There are two copies of P,
both of which get to execute, and whose traces are interleaved together.
That means we can't treat the operand as a set, since that throws away
duplicates.  A bag gives us exactly what we want, though — it keeps
track of cardinality but not order.  (That is, P ⫴ Q is the same process
as Q ⫴ P.)

Note that this is different from the choice operators: P □ P is the same
as □{P} is the same as P.