Add traced processes
I'm really liking this idea of using the process interface to model
basically everything in this refinement checker.  That's worth expanding
on in a blog post.  Here we're using it to keep track of what traces you
follow to get to a subprocess; we'll use that to construct a
counterexample during a refinement check.